Add `cli_install.yml`
Since it is not possible to pre-generate LocalConfigs.php with environment variables or other settings, it should be sensible to let users run 'run.php install' to install a MediaWiki instance and mount the generated LocalConfigs.php.

Note that it is not possible to specify the file's path which the MediaWiki instance should load it — the file must reside inside the installation directory. Otherwise the instance simply does not work.